refactor: move commands to root package
