+++ cmd/oras/root/blob/delete.go
