+++ cmd/oras/root/cmd.go
@@ -10,23 +10,33 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-
-package repository
+package root
 
 import (
 	"github.com/spf13/cobra"
+	"oras.land/oras/cmd/oras/root/blob"
+	"oras.land/oras/cmd/oras/root/manifest"
+	"oras.land/oras/cmd/oras/root/repo"
 )
 
-func Cmd() *cobra.Command {
+func New() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:     "repo [command]",
-		Short:   "Repository operations",
-		Aliases: []string{"repository"},
+		Use:          "oras [command]",
+		SilenceUsage: true,
 	}
-
 	cmd.AddCommand(
-		listCmd(),
-		showTagsCmd(),
+		pullCmd(),
+		pushCmd(),
+		loginCmd(),
+		logoutCmd(),
+		versionCmd(),
+		discoverCmd(),
+		copyCmd(),
+		tagCmd(),
+		attachCmd(),
+		blob.Cmd(),
+		manifest.Cmd(),
+		repo.Cmd(),
 	)
 	return cmd
 }
