+++ cmd/oras/root/logout.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package root
 
 import (
 	"github.com/sirupsen/logrus"
