+++ cmd/oras/root/manifest/delete.go
