+++ cmd/oras/root/repo/cmd.go
@@ -3,9 +3,7 @@ Copyright The ORAS Authors.
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
-
 http://www.apache.org/licenses/LICENSE-2.0
-
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -13,21 +11,22 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package root
+package repo
 
 import (
-	"context"
-	"os"
-	"os/signal"
-
-	"oras.land/oras/cmd/oras/root"
+	"github.com/spf13/cobra"
 )
 
-func main() {
-	ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt)
-	defer cancel()
-	if err := root.New().ExecuteContext(ctx); err != nil {
-		cancel()
-		os.Exit(1)
+func Cmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "repo [command]",
+		Short:   "Repository operations",
+		Aliases: []string{"repository"},
 	}
+
+	cmd.AddCommand(
+		listCmd(),
+		showTagsCmd(),
+	)
+	return cmd
 }
